refactor: change structure app
