+++ src/app.module.ts
@@ -1,16 +1,16 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { APP_GUARD } from '@nestjs/core';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AlbumModule } from './album/album.module';
-import { ArtistModule } from './artist/artist.module';
-import { TrackModule } from './track/track.module';
-import { FavoritesModule } from './favorites/favorites.module';
-import { UserModule } from './user/user.module';
-import { AuthModule } from './auth/auth.module';
+import { AlbumModule } from './modules/album/album.module';
+import { ArtistModule } from './modules/artist/artist.module';
+import { TrackModule } from './modules/track/track.module';
+import { FavoritesModule } from './modules/favorites/favorites.module';
+import { UserModule } from './modules/user/user.module';
+import { AuthModule } from './modules/auth/auth.module';
+import { JwtAuthGuard } from './modules/auth/quards/jwt-auth.guard';
 import { typeormConfig } from './ormconfig';
-import { APP_GUARD } from '@nestjs/core';
-import { JwtAuthGuard } from './auth/quards/jwt-auth.guard';
 
 @Module({
   imports: [
