--- src/auth/auth.module.ts
@@ -1,27 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { JwtModule } from '@nestjs/jwt';
-import { PassportModule } from '@nestjs/passport';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { JwtStrategy } from './strategies/jwt.strategy';
-import { User } from '../user/entities/user.entity';
-import { UserModule } from '../user/user.module';
-import { JWT_SECRET_KEY, TOKEN_EXPIRE_TIME } from '../settings/index';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([User]),
-    JwtModule.register({
-      secret: JWT_SECRET_KEY,
-      signOptions: {
-        expiresIn: TOKEN_EXPIRE_TIME,
-      },
-    }),
-    PassportModule,
-    UserModule,
-  ],
-  controllers: [AuthController],
-  providers: [AuthService, JwtStrategy],
-})
-export class AuthModule {}
