+++ src/modules/favorites/favorites.module.ts
