+++ src/modules/user/entities/user.entity.ts
@@ -1,0 +1,46 @@
+import { Exclude } from 'class-transformer';
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+  VersionColumn,
+} from 'typeorm';
+import { IUser } from '../../../types/index';
+
+@Entity('users')
+export class User implements IUser {
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
+
+  @Column()
+  login: string;
+
+  @VersionColumn()
+  version: number;
+
+  @CreateDateColumn({
+    transformer: {
+      from: (value: Date) => value.getTime(),
+      to: (value: Date) => value,
+    },
+  })
+  createdAt: number;
+
+  @UpdateDateColumn({
+    transformer: {
+      from: (value: Date) => value.getTime(),
+      to: (value: Date) => value,
+    },
+  })
+  updatedAt: number;
+
+  @Exclude()
+  @Column()
+  password: string;
+
+  constructor(userInfo: Partial<User>) {
+    Object.assign(this, userInfo);
+  }
+}
