--- src/track/entities/track.entity.ts
@@ -1,42 +1,0 @@
-import {
-  Column,
-  Entity,
-  JoinColumn,
-  ManyToOne,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
-import { Album } from '../../album/entities/album.entity';
-import { Artist } from '../../artist/entities/artist.entity';
-import { ID_ENTITY_OPTIONS } from '../../settings/index';
-import { getJoinColumnOptions } from '../../utils/index';
-import { ITrack } from '../../types/index';
-
-@Entity('tracks')
-export class Track implements ITrack {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column()
-  name: string;
-
-  @Column({ nullable: true })
-  albumId: string | null;
-
-  @ManyToOne(() => Album, (album) => album.id, ID_ENTITY_OPTIONS)
-  @JoinColumn(getJoinColumnOptions('albumId'))
-  album: string | null;
-
-  @Column({ nullable: true })
-  artistId: string | null;
-
-  @ManyToOne(() => Artist, (artist) => artist.id, ID_ENTITY_OPTIONS)
-  @JoinColumn(getJoinColumnOptions('artistId'))
-  artist: string | null;
-
-  @Column()
-  duration: number;
-
-  constructor(trackInfo: Partial<Track>) {
-    Object.assign(this, trackInfo);
-  }
-}
